v5.01 --- Assignment 5 submission ---

+++ core/src/com/packtpub/libgdx/canyonbunny/game/WorldRenderer.java
@@ -49,7 +49,7 @@ public class WorldRenderer implements Disposable {
 		camera.viewportWidth = (Constants.VIEWPORT_HEIGHT / height) * width;
 		camera.update();
 		cameraGUI.viewportHeight = Constants.VIEWPORT_GUI_HEIGHT;
-		cameraGUI.viewportWidth = (Constants.VIEWPORT_GUI_HEIGHT / (float)height) * (float)width;
+		cameraGUI.viewportWidth = (Constants.VIEWPORT_GUI_HEIGHT / height) * width;
 		cameraGUI.position.set(cameraGUI.viewportWidth/2,cameraGUI.viewportHeight/2,0);
 		cameraGUI.update();
 	}
